requirements.txt
ajsonrpc==1.2.0
anyio==4.0.0
beautifulsoup4==4.12.2
bottle==0.12.25
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.5
marshmallow==3.20.1
matplotlib==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformio==6.1.11
pyelftools==0.29
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.31.1
tabulate==0.9.0
tk==0.1.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
wsproto==1.2.0
